+++ Navtrack.Listener/Helpers/GpsUtil.cs
@@ -6,14 +6,14 @@ namespace Navtrack.Listener.Helpers;
 
 public static class GpsUtil
 {
-    public static bool IsValidLatitude(decimal latitude) => latitude is >= -90 and <= 90;
+    public static bool IsValidLatitude(double latitude) => latitude is >= -90 and <= 90;
 
-    public static bool IsValidLongitude(decimal longitude) => longitude is >= -180 and <= 180;
+    public static bool IsValidLongitude(double longitude) => longitude is >= -180 and <= 180;
 
-    public static decimal ConvertDmmLatToDecimal(string point, string cardinalDirection) =>
+    public static double ConvertDmmLatToDecimal(string point, string cardinalDirection) =>
         ConvertDmmToDecimal(@"(\d+)(\d\d.\d+)", point, cardinalDirection);
         
-    public static decimal ConvertDmmLongToDecimal(string point, string cardinalDirection) =>
+    public static double ConvertDmmLongToDecimal(string point, string cardinalDirection) =>
         ConvertDmmToDecimal(@"(\d+)(\d\d.\d+)", point, cardinalDirection);
 
     public static decimal ConvertDmsToDecimal(string regExPattern, string point, string cardinalDirection)
@@ -28,36 +28,36 @@ public static class GpsUtil
         return Math.Round((degrees + minutes + seconds) * multiplier, 6, MidpointRounding.ToZero);
     }
 
-    public static decimal ConvertDmmToDecimal(string regExPattern, string point, string cardinalDirection)
+    public static double ConvertDmmToDecimal(string regExPattern, string point, string cardinalDirection)
     {
         MatchCollection matchCollection = new Regex(regExPattern).Matches(point);
 
         int multiplier = cardinalDirection == "S" || cardinalDirection == "W" ? -1 : 1;
-        decimal degrees = decimal.Parse(matchCollection[0].Groups[1].Value);
-        decimal minutes = decimal.Parse(matchCollection[0].Groups[2].Value) / 60;
+        double degrees = double.Parse(matchCollection[0].Groups[1].Value);
+        double minutes = double.Parse(matchCollection[0].Groups[2].Value) / 60;
 
         return Math.Round((degrees + minutes) * multiplier, 6, MidpointRounding.ToZero);
     }
 
-    public static decimal ConvertDmmToDecimal(decimal degrees, decimal minutes, CardinalPoint cardinalPoint)
+    public static double ConvertDmmToDecimal(double degrees, double minutes, CardinalPoint cardinalPoint)
     {
         int multiplier = cardinalPoint == CardinalPoint.South || cardinalPoint == CardinalPoint.West ? -1 : 1;
           
         return Math.Round((degrees + minutes) * multiplier, 6, MidpointRounding.ToZero);
     }
 
-    public static decimal ConvertDdmToDecimal(decimal degrees, decimal minutes, CardinalPoint cardinalPoint)
+    public static double ConvertDdmToDecimal(double degrees, double minutes, CardinalPoint cardinalPoint)
     {
         int multiplier = cardinalPoint == CardinalPoint.South || cardinalPoint == CardinalPoint.West ? -1 : 1;
           
         return Math.Round((degrees + minutes/60) * multiplier, 6, MidpointRounding.ToZero);
     }
 
-    public static decimal ConvertStringToDecimal(string value, string cardinalDirection)
+    public static double ConvertStringToDecimal(string value, string cardinalDirection)
     {
         int multiplier = cardinalDirection == "S" || cardinalDirection == "W" ? -1 : 1;
 
-        decimal result = decimal.Parse(value) * multiplier;
+        double result = double.Parse(value) * multiplier;
 
         return result;
     }
