+++ Navtrack.Listener/Models/Location.cs
@@ -20,15 +20,15 @@ public class Location
     }
 
     public Device Device { get; set; }
-    public decimal Latitude { get; set; }
-    public decimal Longitude { get; set; }
+    public double Latitude { get; set; }
+    public double Longitude { get; set; }
     public DateTime DateTime { get; set; }
     public bool? PositionStatus { get; set; }
-    public decimal? Speed { get; set; }
-    public decimal? Heading { get; set; }
-    public decimal? Altitude { get; set; }
+    public float? Speed { get; set; }
+    public float? Heading { get; set; }
+    public float? Altitude { get; set; }
     public short? Satellites { get; set; }
-    public decimal? HDOP { get; set; }
+    public float? HDOP { get; set; }
     public short? GsmSignal { get; set; }
     public double? Odometer { get; set; }
