+++ Navtrack.Listener/Protocols/Coban/CobanMessageHandler.cs
@@ -66,11 +66,11 @@ public class CobanMessageHandler : BaseMessageHandler<CobanProtocol>
                 input.DataMessage.CommaSplit[8]),
             Longitude = GpsUtil.ConvertDmmLongToDecimal(input.DataMessage.CommaSplit[9],
                 input.DataMessage.CommaSplit[10]),
-            Speed = SpeedUtil.KnotsToKph(input.DataMessage.CommaSplit.Get<decimal>(11)),
+            Speed = SpeedUtil.KnotsToKph(input.DataMessage.CommaSplit.Get<float>(11)),
             Heading = input.DataMessage.CommaSplit.Get<string>(12) != "1"
-                ? input.DataMessage.CommaSplit.Get<decimal?>(12)
+                ? input.DataMessage.CommaSplit.Get<float?>(12)
                 : null,
-            Altitude = input.DataMessage.CommaSplit.Get<decimal?>(13),
+            Altitude = input.DataMessage.CommaSplit.Get<float?>(13),
         };
 
         return location;
